Implement game() function and multiple rounds
The game function contains player- and computer score. A for-loop allows
the game to be played for 5 rounds, where each round contains the
selections and comparisons. A round will be replayed every time there is
a tie. The final score is displayed along with who the winner is.

The playRound() function has been changed to take the round number as a
parameter and include it in the result message for each round.