Add button functionality
Queries, and adds event listeners, to buttons for 'rock', 'paper' and
'scissors'. Listeners call the playOneRound function where the button id
is sent in as the player choice.

game() is removed, with the variables playerScore and computerScore
global instead of local. The rest of the code is incorporated into
playOneRound(). playOneRound() now takes playerSelection as input,
checks if the game is already over with checkGameOver() and then plays
the round. Score is updated in updateScore() and checks if the game is
over with checkGameOver(). It then displays the game over message if the
game is over.